+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/exceptions/RequestException.kt
@@ -9,11 +9,12 @@ fun newRequestException(
     message: String? = null,
     cause: Throwable? = null
 ) = when (response.description) {
-    "Bad Request: reply message not found" -> ReplyMessageNotFound(response, plainAnswer, message, cause)
-    else -> RequestException(response, plainAnswer, message, cause)
+    "Bad Request: reply message not found" -> ReplyMessageNotFoundException(response, plainAnswer, message, cause)
+    "Unauthorized" -> UnauthorizedException(response, plainAnswer, message, cause)
+    else -> CommonRequestException(response, plainAnswer, message, cause)
 }
 
-open class RequestException internal constructor(
+sealed class RequestException constructor(
     val response: Response<*>,
     val plainAnswer: String,
     message: String? = null,
@@ -22,3 +23,18 @@ open class RequestException internal constructor(
     message,
     cause
 )
+
+class CommonRequestException(response: Response<*>, plainAnswer: String, message: String?, cause: Throwable?) :
+    RequestException(response, plainAnswer, message, cause)
+
+class UnauthorizedException(response: Response<*>, plainAnswer: String, message: String?, cause: Throwable?) :
+    RequestException(response, plainAnswer, message, cause)
+
+class ReplyMessageNotFoundException(response: Response<*>, plainAnswer: String, message: String?, cause: Throwable?) :
+    RequestException(response, plainAnswer, message, cause)
+
+@Deprecated(
+    "Replaced by ReplyMessageNotFoundException",
+    ReplaceWith("ReplyMessageNotFoundException", "com.github.insanusmokrassar.TelegramBotAPI.bot.exceptions.ReplyMessageNotFoundException")
+)
+typealias ReplyMessageNotFound = ReplyMessageNotFoundException